requirements.txt
strands-agents==1.19.0
strands-agents-tools==0.2.17
PyGithub==2.5.0
pytest==8.3.4
